rs-recognize/requirements.txt
functions-framework==3.*
tensorflow
google-cloud-storage
numpy